# Conflicts:
#	API-Paris/src/test/java/com/paris/api/ApiParisApplicationTests.java
#	API-Paris/target/test-classes/com/paris/api/ApiParisApplicationTests.class

+++ API-Paris/src/main/java/com/paris/api/services/ParieService.java
@@ -4,7 +4,6 @@ import com.paris.api.models.ParieModel;
 import com.paris.api.repository.ParieRepository;
 import org.apache.commons.math3.util.Precision;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
 
 import java.io.Serializable;
 import java.util.List;
@@ -19,31 +18,46 @@ public class ParieService implements Serializable {
         ParieService.repository = repository;
     }
 
-    public ParieModel findByID(Long id) {
-        return repository.findById(id).orElse(new ParieModel(null, null, 0,0,0));
-    }
-    @GetMapping("/all")
+
     public static List<ParieModel> all() {
         return repository.findAll();
     }
-
-    public ParieModel createParie(ParieModel model) {
-        return repository.save(model);
+    public ParieModel findByID(Long id) {
+        return repository.findById(id).orElse(new ParieModel(null, null, 0,0,0));
     }
 
-    public String deleteByID(Long id){
-        repository.deleteById(id);
-        return "Parieur supprimer";
+
+    public boolean delByID(Long id){
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            return true;
+        }else{
+            return false;
+        }
+
     }
 
-    public Boolean addParie(Long idMatch, Float scoreEqip1, Float scoreEqip2){
-        ParieModel newParie = new ParieModel(null, null, 0,0,0);
-        newParie.setIdMatch(idMatch);
-        newParie.setCoteA(calculChances(scoreEqip1, scoreEqip2, "A"));
-        newParie.setCoteB(calculChances(scoreEqip1, scoreEqip2, "B"));
-        newParie.setCoteN(calculChances(scoreEqip1, scoreEqip2, "N"));
-        repository.save(newParie);
-        return true;
+    public ParieModel addParie(Long idMatch, Float scoreEqip1, Float scoreEqip2){
+        List<ParieModel> model = repository.findByidMatch(idMatch);
+        if (model.isEmpty()){
+            ParieModel newParie = new ParieModel(null, null, 0,0,0);
+            newParie.setIdMatch(idMatch);
+            newParie.setCoteA(calculChances(scoreEqip1, scoreEqip2, "A"));
+            newParie.setCoteB(calculChances(scoreEqip1, scoreEqip2, "B"));
+            newParie.setCoteN(calculChances(scoreEqip1, scoreEqip2, "N"));
+            return repository.save(newParie);}
+        else {
+            return new ParieModel();
+        }
+    }
+
+    public ParieModel createPari(ParieModel pari) {
+        List<ParieModel> model = repository.findByidMatch(pari.getIdMatch());
+        if (model.isEmpty()) {
+            return repository.save(pari);
+        }else{
+            return new ParieModel();
+        }
     }
 
     private static Float calculChances (Float score1, Float score2, String call) {
